Fixed bad syntax in registry test.
`setup` blocks should return a tuple with `:ok` and a *keyword list* of the
test context.

The return values of `GenServer.call` is a tuple with
`:ok` and the return value specified in your `handle_call` implementation. The
return value is *not* a tuple with `:reply`. That tuple is expected internally
by GenServer, but is not exposed to the client.