Monitor Bucket processes for crashes from within the registry.
Whenever we add a bucket to the registry we now monitor it, getting back
a unique `ref` for that bucket.

We use those refs to build a map with refs pointing to the names of buckets.

When a bucket crashes (which we detect inside the `handle_info` GenServer
callback), we look up the bucket's name by its `ref` (which is passed along in
the `:DOWN` message) and remove that bucket name from the registry's state.

Right now our registry process is linked to every bucket process, which means
that the registry will crash if any bucket crashes, so we'll need to resolve
this.